Allow custom values for the hidden input
Previously, the value of the hidden input element was always the label
of the selected item. Especially in non-trival cases where the items are
complex objects, it might not be desirable to have the label, but rather
the full item or its id are more interesting.

This change is backwards compatible in that it adds a new property
`getHiddenInputValue` which defaults to the previous behaviour to use
the label.